Use _NET_INPUT_AREAS to decide GtkEntryCompletion popup positioning.
The matchbox version on Sardine has a new X property for the root window
called _NET_INPUT_AREAS. It's supposed to contain an array of rectangles
which are reserved for user interaction and should never be obscured
(Note: the current matchbox impl. only seems to supports one input area at the time,
I'm assuming this is a bug and have developed the general case. When Tapani
is back I'll poke him about this).

I've modified the GtkEntryCompletion popup position code to:
* Get a list of the available input areas.
* Compute the intersections with the rectangle composed with
  X = entry's X, Y = 0, WIDTH = entry's WIDTH, HEIGHT = monitors' HEIGHT
* From that data gather which one is the nearest intersecting area above
  and below the entry.
* Based on the distances to those areas, decide where to show the popup.

Known problems: I don't really know what to do when there's no space either
way to show the popup.


git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk@12940 50eb543b-83fa-0310-9bb7-f7cbbee4ab39