Make dialogs request the full width of the screen
When a dialog is created, its allocation when the first expose event
is called will be its natural width. Then later a new size allocation
is set and the dialog is drawn again with the full width of the
screen.

If the content of the dialog depends on its width (wrapping labels for
example) then the resizing is visible and is very ugly.

Fixes: NB#155392 (GtkDialog are not made full width)